requirements.txt
certifi==2018.4.16
chardet==3.0.4
Cython==0.26.1
docutils==0.14
idna==2.7
Kivy==1.10.1.dev0
Kivy-Garden==0.1.4
pycryptodome==3.6.1
Pygments==2.2.0
requests==2.19.1
urllib3==1.23
